pyproject.toml
[tool.poetry]
name = "avpz-backend"
version = "0.1.0"
description = ""
authors = ["RuslanUC <dev_ruslan_uc@protonmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.31.0"
fastapi = "^0.115.0"
tortoise-orm = {extras = ["accel", "aiomysql"], version = "^0.21.6"}
aerich = "^0.7.2"
pydantic = {extras = ["email"], version = "^2.9.2"}
bcrypt = "^4.2.0"
httpx = "^0.27.2"
pydantic-extra-types = {extras = ["phonenumbers"], version = "^2.9.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.24.0"
asgi-lifespan = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov-report=xml --cov-append --cov=hhb"

[tool.coverage.run]
data_file = "coverage.coverage"